refactor: extract LensForm from LensViewRenderer
